+++ ntest_timeout/tests/timeout.rs
@@ -16,3 +16,11 @@ fn timeout() {
     let fifty_millis = time::Duration::from_millis(50);
     thread::sleep(fifty_millis);
 }
+
+#[test]
+#[timeout(100)]
+fn timeout_with_result() -> Result<(), String> {
+    let ten_millis = time::Duration::from_millis(10);
+    thread::sleep(ten_millis);
+    Ok(())
+}